requirements.txt
pillow 
requests
svgwrite
markdown-it-py[plugins]